+++ src/commands/battle/partyRemove.js
@@ -17,8 +17,8 @@ const partyRemove = async (user, option) => {
         const index = position - 1;
         
         // check if position is valid
-        if (index < 0 || index >= trainer.data.party.length) {
-            return { send: null, err: `Invalid position! Must be between 1 and ${trainer.data.party.length}.` };
+        if (index < 0 || index >= trainer.data.party.pokemonIds.length) {
+            return { send: null, err: `Invalid position! Must be between 1 and ${trainer.data.party.pokemonIds.length}.` };
         }
 
         // check if pokemon exists in position