+++ src/services/rateLimit.js
@@ -3,7 +3,7 @@ const userLastRequest = {}
 // TODO: cleanup userLastRequest to remove old entries
 /**
  * Given a user id, check to see when the user last made a successful request.
- * If the user has not made a request in the last second, return true and update 
+ * If the user has not made a request in the last 3/4ths second, return true and update 
  * the last request time. Else, return false.
  * @param {String} userId 
  */
@@ -13,7 +13,7 @@ const checkRateLimit = (userId) => {
         userLastRequest[userId] = now;
         return true;
     }
-    if (now - userLastRequest[userId] > 1000) {
+    if (now - userLastRequest[userId] > 750) {
         userLastRequest[userId] = now;
         return true;
     }